requirements.txt
fastapi 
uvicorn 
gunicorn 
pydantic
python-dotenv
requests 
beautifulsoup4 
langchain 
langchain-openai
streamlit 
 